+++ update-client-hub/modules/source/update-client-source/arm_uc_source.h
@@ -20,7 +20,7 @@
 #define __ARM_UPDATE_SOURCE_H__
 
 #include "update-client-common/arm_uc_error.h"
-#include "update-client-common/arm_uc_types.h"
+#include "update-client-common/arm_uc_types_internal.h"
 
 #include <stdint.h>
 
@@ -59,7 +59,7 @@ typedef enum _ARM_SOURCE_EVENT {
 /**
  * @brief Prototype for event handler.
  */
-typedef void (*ARM_SOURCE_SignalEvent_t)(uint32_t event);
+typedef void (*ARM_SOURCE_SignalEvent_t)(uintptr_t event);
 
 /**
  * @brief Structure definition holding API function pointers.
